NativeMethods: undo partial classes
The GVFS/GVFS.Common/NativeMethods.cs file was split into two
and GVFS/GVFS.Common/NativeMethods.Shared.cs was created in
commit 785ccb4a67281d0eb10f19cb53e6431b62b88555, in order to
share the ThrowLastWin32Exception() and GetFinalPathName()
methods and some native Win32 enums and system call wrappers
with the GVFS.Hooks project.

Given that there is no longer a corresponding Scalar version of
the GVFS.Hooks project, there is no need to split out certain
(static) methods from the others in any classes.  (The GVFS.Hooks
project was removed in commit
f12effc371d7b1457ab58c6d44959632404e5998 in PR #23.)

Thus we merge the two methods, file attribute and access enums,
and system call wrappers into NativeMethods.cs and remove the
NativeMethods.Shared.cs file.
